refactor: add shared streaming-memory tests with v4/v5+ model support
