+++ packages/memory/integration-tests-v5/src/streaming-memory.test.ts
@@ -18,136 +18,8 @@ import { JSDOM } from 'jsdom';
 import { describe, expect, it, beforeAll, afterAll } from 'vitest';
 import { memory, weatherAgent } from './mastra/agents/weather';
 import { weatherTool } from './mastra/tools/weather';
-// Helper to find an available port
-async function getAvailablePort(): Promise<number> {
-  return new Promise((resolve, reject) => {
-    const server = createServer();
-    server.listen(0, () => {
-      const { port } = server.address() as { port: number };
-      server.close(() => resolve(port));
-    });
-    server.on('error', reject);
-  });
-}
-
-// Set up JSDOM environment for React testing
-const dom = new JSDOM('<!doctype html><html><body></body></html>', {
-  url: 'http://localhost',
-  pretendToBeVisual: true,
-  resources: 'usable',
-});
-// @ts-ignore - JSDOM types don't match exactly but this works for testing
-global.window = dom.window;
-global.document = dom.window.document;
-Object.defineProperty(global, 'navigator', {
-  value: dom.window.navigator,
-  writable: false,
-});
-global.fetch = global.fetch || fetch;
 
 describe('Memory Streaming Tests', () => {
-  it('should handle multiple tool calls in memory thread history', async () => {
-    // Create agent with memory and tools
-    const agent = new Agent({
-      id: 'test-agent',
-      name: 'test',
-      instructions:
-        'You are a weather agent. When asked about weather in any city, use the get_weather tool with the city name as the postal code. Respond in a pirate accent and dont use the degrees symbol, print the word degrees when needed.',
-      model: openai('gpt-4o'),
-      memory,
-      tools: { get_weather: weatherTool },
-    });
-
-    const threadId = randomUUID();
-    const resourceId = randomUUID();
-
-    // First weather check
-    const stream1 = await agent.stream('what is the weather in LA?', {
-      threadId,
-      resourceId,
-    });
-
-    // Collect first stream
-    const chunks1: string[] = [];
-    for await (const chunk of stream1.fullStream) {
-      if (chunk.type === `text-delta`) {
-        chunks1.push(chunk.payload.text);
-      }
-    }
-    const response1 = chunks1.join('');
-
-    expect(chunks1.length).toBeGreaterThan(0);
-    expect(response1).toContain('70 degrees');
-
-    // Second weather check
-    const stream2 = toAISdkStream(
-      await agent.stream('what is the weather in Seattle?', {
-        threadId,
-        resourceId,
-      }),
-      { from: 'agent' },
-    );
-
-    // Collect second stream
-    const chunks2: string[] = [];
-
-    for await (const chunk of stream2) {
-      if (chunk.type === `text-delta`) {
-        chunks2.push(chunk.delta);
-      }
-    }
-    const response2 = chunks2.join('');
-
-    expect(chunks2.length).toBeGreaterThan(0);
-    expect(response2).toContain('Seattle');
-    expect(response2).toContain('70 degrees');
-  });
-
-  it('should use custom mastra ID generator for messages in memory', async () => {
-    const agent = new Agent({
-      id: 'test-msg-id-agent',
-      name: 'test-msg-id',
-      instructions: 'you are a helpful assistant.',
-      model: openai('gpt-4o'),
-      memory,
-    });
-
-    const threadId = randomUUID();
-    const resourceId = 'test-resource-msg-id';
-    const customIds: UUID[] = [];
-
-    new Mastra({
-      idGenerator: () => {
-        const id = randomUUID();
-        customIds.push(id);
-        return id;
-      },
-      agents: {
-        agent: agent,
-      },
-    });
-
-    await agent.generate('Hello, world!', {
-      threadId,
-      resourceId,
-    });
-
-    const agentMemory = (await agent.getMemory())!;
-    const { messages } = await agentMemory.recall({ threadId });
-
-    console.log('Custom IDs: ', customIds);
-    console.log('Messages: ', messages);
-
-    expect(messages).toHaveLength(2);
-    expect(messages.length).toBeLessThan(customIds.length);
-    for (const message of messages) {
-      if (!(`id` in message)) {
-        throw new Error(`Expected message.id`);
-      }
-      expect(customIds).contains(message.id);
-    }
-  });
-
   describe('should stream via useChat after tool call', () => {
     let mastraServer: ReturnType<typeof spawn>;
     let port: number;
