+++ packages/memory/integration-tests/package.json
@@ -34,11 +34,11 @@
     "dev": "mastra dev"
   },
   "dependencies": {
+    "@ai-sdk/anthropic-v5": "npm:@ai-sdk/anthropic@^2.0.0",
+    "@ai-sdk/anthropic-v6": "npm:@ai-sdk/anthropic@latest",
     "@ai-sdk/google": "^1.2.19",
     "@ai-sdk/google-v5": "npm:@ai-sdk/google@^2.0.0",
     "@ai-sdk/google-v6": "npm:@ai-sdk/google@latest",
-    "@ai-sdk/anthropic-v5": "npm:@ai-sdk/anthropic@^2.0.0",
-    "@ai-sdk/anthropic-v6": "npm:@ai-sdk/anthropic@latest",
     "@ai-sdk/openai": "^1.3.0",
     "@ai-sdk/openai-v5": "npm:@ai-sdk/openai@^2.0.0",
     "@ai-sdk/openai-v6": "npm:@ai-sdk/openai@latest",
@@ -58,6 +58,7 @@
     "@faker-js/faker": "^9.9.0",
     "@mastra/core": "*",
     "@testing-library/react": "^16.3.0",
+    "@types/jsdom": "^27.0.0",
     "@types/json-schema": "^7.0.15",
     "@types/node": "22.13.17",
     "ai": "^4.2.2",
